requirement.txt
Flask==2.3.3
Pillow==10.0.0
opencv-python==4.8.1.78
numpy==1.24.3
reportlab==4.0.4
requests==2.31.0
transformers==4.35.0  # Optional - if you want to use BLIP later
torch==2.1.0  # Optional - if you want to use BLIP later